Revert "Change floats to doubles"
This reverts commit 5bf316bc77c8ce25210f1387467c99a1b1b9ae8b.

+++ ros1/src/data/march_realsense_reader/include/pointcloud_processor/parameter_determiner.h
@@ -148,17 +148,17 @@ protected:
     // Read all relevant parameters
     int hull_dimension;
     int number_of_optional_foot_locations;
-    double min_x_stairs;
-    double max_x_stairs;
-    double min_z_stairs;
-    double max_z_stairs;
-    double y_location;
+    float min_x_stairs;
+    float max_x_stairs;
+    float min_z_stairs;
+    float max_z_stairs;
+    float y_location;
     double x_flat;
     double z_flat;
     double x_steep;
     double z_steep;
-    double min_search_area;
-    double max_search_area;
+    float min_search_area;
+    float max_search_area;
     double max_distance_to_line;
     bool general_most_desirable_location_is_mid;
     bool general_most_desirable_location_is_small;
